Added policy methods to authorize if user is allowed to unshare namespace with groups, added translations, added tests

+++ app/services/namespaces/group_unshare_service.rb
@@ -3,6 +3,7 @@
 module Namespaces
   # Service used to Destroy NamespaceGroupLinks
   class GroupUnshareService < BaseService
+    NamespaceGroupUnshareError = Class.new(StandardError)
     attr_accessor :group_id, :namespace, :max_group_access_role
 
     def initialize(user, group_id, namespace)
@@ -12,13 +13,16 @@ module Namespaces
     end
 
     def execute
-      # authorize! namespace, to: :unshare_namespace_with_group?
+      authorize! namespace, to: :unshare_namespace_with_group?
 
       namespace_group_link = NamespaceGroupLink.find_by(group_id:, namespace:)
 
-      namespace_group_link.destroy
+      raise NamespaceGroupUnshareError, 'Namespace to group link does not exist' if namespace_group_link.nil?
 
-      namespace_group_link
+      namespace_group_link.destroy
+    rescue Namespaces::GroupUnshareService::NamespaceGroupUnshareError => e
+      @namespace.errors.add(:base, e.message)
+      false
     end
   end
 end
